Move Fax to end of phone number priorities
The phoneNumberFields is used as a priority order for choosing a possible phone number to display for a person (when only one number is desired).

The fax number is now moved to the end of the list.  There is no good reason to display a fax number for a person, when we only display a single number for them.

Updates #49